+++ packages-core/lint/src/lint.test.ts
@@ -1,21 +1,23 @@
-import anyTest, {TestInterface} from 'ava';
-import {join} from 'path';
-import {readFile} from 'fs';
-import {Linter} from 'eslint';
-import {rules as availableTypeScriptRules} from '@typescript-eslint/eslint-plugin';
+import anyTest, {TestInterface, ExecutionContext} from 'ava';
+import * as path from 'path';
+import * as fs from 'fs';
+import * as eslint from 'eslint';
+import * as eslintPlugin from '@typescript-eslint/eslint-plugin';
 
-interface IESLintConfig extends Linter.Config {
-    overrides: Array<Linter.Config>,
+interface IESLintConfig extends eslint.Linter.Config {
+    overrides: Array<eslint.Linter.Config>,
 }
 
-const test = anyTest as TestInterface<{
+interface ITextContext {
     config: IESLintConfig,
-}>;
+}
+
+const test = anyTest as TestInterface<ITextContext>;
 
 test.beforeEach(async (t) => {
     const json: string = await new Promise((resolve, reject) => {
-        const packageJson = join(__dirname, '..', '.eslintrc.json');
-        readFile(packageJson, 'utf8', (error, string) => {
+        const packageJson = path.join(__dirname, '..', '.eslintrc.json');
+        fs.readFile(packageJson, 'utf8', (error, string) => {
             if (error) {
                 reject(error);
             } else {
@@ -27,17 +29,29 @@ test.beforeEach(async (t) => {
 });
 
 test('should be valid configuration', (t) => {
-    const linter = new Linter();
+    const linter = new eslint.Linter();
     t.truthy(linter.verify('', t.context.config));
 });
 
 {
-    const availableESLintRules = new Linter().getRules();
-    for (const [ruleName] of availableESLintRules) {
-        test(`should cover "${ruleName}"`, (t) => {
-            const {rules = {}} = t.context.config;
-            t.truthy(rules[ruleName], `${ruleName} is not covered`);
-        });
+    const availableESLintRules = new eslint.Linter().getRules();
+    for (const [ruleName, rule] of availableESLintRules) {
+        const url = `https://eslint.org/docs/rules/${ruleName}`;
+        if (rule.meta && rule.meta.deprecated) {
+            test(`should not cover deprecated "${ruleName}"`, (t) => {
+                const {rules = {}} = t.context.config;
+                t.is(
+                    rules[ruleName],
+                    undefined,
+                    `${ruleName} is deprecated.\n${url}`,
+                );
+            });
+        } else {
+            test(`should cover "${ruleName}"`, (t) => {
+                const {rules = {}} = t.context.config;
+                t.truthy(rules[ruleName], `${ruleName} is not covered.\n${url}`);
+            });
+        }
     }
     test('should have supported rules', (t) => {
         for (const rule of Object.keys(t.context.config.rules || {})) {
@@ -47,60 +61,66 @@ test('should be valid configuration', (t) => {
 }
 
 {
+    const availableTypeScriptRules = new Map(Object.entries(eslintPlugin.rules));
     const prefix = '@typescript-eslint';
-    const availableTypeScriptRuleNames = new Set(Object.keys(availableTypeScriptRules).map((name) => `${prefix}/${name}`));
     // https://github.com/typescript-eslint/typescript-eslint/issues/101
     const rulesToBeIgnored = new Set<string>([
-        '@typescript-eslint/await-thenable',
-        '@typescript-eslint/no-floating-promises',
-        '@typescript-eslint/no-for-in-array',
-        '@typescript-eslint/no-unnecessary-qualifier',
-        '@typescript-eslint/no-unnecessary-type-assertion',
-        '@typescript-eslint/prefer-includes',
-        '@typescript-eslint/prefer-regexp-exec',
-        '@typescript-eslint/prefer-string-starts-ends-with',
-        '@typescript-eslint/promise-function-async',
-        '@typescript-eslint/require-array-sort-compare',
-        '@typescript-eslint/restrict-plus-operands',
-        '@typescript-eslint/unbound-method',
+        'await-thenable',
+        'no-floating-promises',
+        'no-for-in-array',
+        'no-unnecessary-qualifier',
+        'no-unnecessary-type-assertion',
+        'prefer-includes',
+        'prefer-regexp-exec',
+        'prefer-string-starts-ends-with',
+        'prefer-readonly',
+        'promise-function-async',
+        'require-array-sort-compare',
+        'restrict-plus-operands',
+        'strict-boolean-expressions',
+        'unbound-method',
     ]);
-    for (const ruleName of availableTypeScriptRuleNames) {
+    const getTypeScriptRules = (
+        t: ExecutionContext<ITextContext>,
+    ): NonNullable<eslint.Linter.Config['rules']> => {
+        const typescriptConfig = t.context.config.overrides
+        .find((override) => override.parser === `${prefix}/parser`) || {};
+        return typescriptConfig.rules || {};
+    };
+    for (const [ruleName, rule] of availableTypeScriptRules) {
+        const url = `https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/${ruleName}.md`;
         if (rulesToBeIgnored.has(ruleName)) {
-            test(`should ignore "${ruleName}"`, (t) => {
-                const typescriptConfig = t.context.config.overrides
-                .find((override) => override.parser === `${prefix}/parser`);
-                if (!typescriptConfig) {
-                    t.truthy(typescriptConfig);
-                    return;
-                }
-                const {rules = {}} = typescriptConfig;
-                const config = rules[ruleName];
-                t.falsy(config, `${ruleName} should be ignored`);
+            test(`should ignore "${prefix}/${ruleName}"`, (t) => {
+                const rules = getTypeScriptRules(t);
+                const config = rules[`${prefix}/${ruleName}`];
+                t.falsy(config, `${prefix}/${ruleName} should be ignored.\n${url}`);
+            });
+        } else if (rule.meta && rule.meta.deprecated) {
+            test(`should not cover deprecated "${prefix}/${ruleName}"`, (t) => {
+                const rules = getTypeScriptRules(t);
+                const config = rules[`${prefix}/${ruleName}`];
+                t.is(
+                    config,
+                    undefined,
+                    `${prefix}/${ruleName} is deprecated.\n${url}`,
+                );
             });
         } else {
-            test(`should cover "${ruleName}"`, (t) => {
-                const typescriptConfig = t.context.config.overrides
-                .find((override) => override.parser === `${prefix}/parser`);
-                if (!typescriptConfig) {
-                    t.truthy(typescriptConfig);
-                    return;
-                }
-                const {rules = {}} = typescriptConfig;
-                const config = rules[ruleName];
-                t.truthy(config, `${ruleName} is not covered`);
+            test(`should cover "${prefix}/${ruleName}"`, (t) => {
+                const rules = getTypeScriptRules(t);
+                const config = rules[`${prefix}/${ruleName}`];
+                t.truthy(config, `${prefix}/${ruleName} is not covered.\n${url}`);
             });
         }
     }
     test('should have supported typescript-eslint rules', (t) => {
-        const typescriptConfig = t.context.config.overrides
-        .find((override) => override.parser === `${prefix}/parser`);
-        if (!typescriptConfig) {
-            t.truthy(typescriptConfig);
-            return;
-        }
-        const {rules = {}} = typescriptConfig;
-        for (const rule of Object.keys(rules).filter((name) => name.startsWith(prefix))) {
-            t.true(availableTypeScriptRuleNames.has(rule), `${rule} is not supported`);
+        const rules = getTypeScriptRules(t);
+        const prefixLength = prefix.length + 1;
+        for (const ruleName of Object.keys(rules).filter((name) => name.startsWith(prefix))) {
+            t.true(
+                availableTypeScriptRules.has(ruleName.slice(prefixLength)),
+                `${ruleName} is not supported`,
+            );
         }
     });
 }