+++ packages-core/lint/src/types.d.ts
@@ -1,4 +1,7 @@
 declare module '@typescript-eslint/eslint-plugin' {
-    export const rules: {};
+    import * as eslint from 'eslint';
+    export const rules: {
+        [name: string]: eslint.Rule.RuleModule,
+    };
     export const configs: {};
 }