+++ packages-node/node-net/src/index.ts
@@ -1,3 +1,5 @@
 export * from './listen';
 export * from './request';
 export * from './httpGet';
+export * from './closeServers';
+export * from './getBaseURL';