+++ packages-core/cookie/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.12.0](https://github.com/nlibjs/nlibjs/compare/v3.11.2...v3.12.0) (2019-07-11)
+
+
+### Bug Fixes
+
+* **cookie:** getCookieStringFor doesn't return value if path is nonempty ([201ea76](https://github.com/nlibjs/nlibjs/commit/201ea76))
+
+
+
+
+
 ## [3.11.2](https://github.com/nlibjs/nlibjs/compare/v3.11.1...v3.11.2) (2019-07-09)
 
 **Note:** Version bump only for package @nlib/cookie