+++ packages-core/number/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.12.0](https://github.com/nlibjs/nlibjs/compare/v3.11.2...v3.12.0) (2019-07-11)
+
+**Note:** Version bump only for package @nlib/number
+
+
+
+
+
 # [3.10.0](https://github.com/nlibjs/nlibjs/compare/v3.9.3...v3.10.0) (2019-06-05)
 
 **Note:** Version bump only for package @nlib/number